refactor(permission): 将权限检查逻辑移动到comm包并清理无用导入
重构权限检查逻辑，将其从cb包移动到comm包以提升代码组织性
移除cb.go中未使用的导入和重复的权限检查代码
