+++ pkg/comm/permission.go
@@ -1,4 +1,4 @@
-package cb
+package comm
 
 import (
 	"context"
